settings and email changes

+++ triage_app/routes/task.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from .. import schemas
 from sqlalchemy.orm import Session
 from ..dependencies import get_db
-from ..crud import create_task, delete_task, update_task, decode_token, get_task_by_filter, get_tasks
+from ..crud import create_task, delete_task, update_task, decode_token, get_task_by_filter, get_tasks, get_role
 from fastapi.responses import JSONResponse
 
 
@@ -10,6 +10,8 @@ router = APIRouter(prefix='/task')
 
 @router.post("/create", response_model=schemas.Task)
 def task_create(task: schemas.TaskCreate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_role(db=db, agent_id=agent_data.agent_id, role='task.create'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")    
     return create_task(db=db, task=task)
 
 
@@ -26,6 +28,8 @@ def get_all_tasks(db: Session = Depends(get_db), agent_data: schemas.TokenData =
 
 @router.put("/put/{task_id}", response_model=schemas.Task)
 def task_update(task_id: int, updates: schemas.TaskUpdate, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_role(db=db, agent_id=agent_data.agent_id, role='task.edit'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     task = update_task(db, task_id, updates)
     if not task:
         raise HTTPException(status_code=400, detail=f'Task with id {task_id} not found')
@@ -34,6 +38,8 @@ def task_update(task_id: int, updates: schemas.TaskUpdate, db: Session = Depends
 
 @router.delete("/delete/{task_id}")
 def task_delete(task_id: int, db: Session = Depends(get_db), agent_data: schemas.TokenData = Depends(decode_token)):
+    if not get_role(db=db, agent_id=agent_data.agent_id, role='task.delete'):
+        raise HTTPException(status_code=403, detail="Access denied: You do not have permission to access this resource")
     status = delete_task(db, task_id)
     if not status:
         raise HTTPException(status_code=400, detail=f'Task with id {task_id} not found')
